Remove precision 16 to avoid NaNs in batch multiplication in Agents Module

+++ agents_module.py
@@ -57,9 +57,9 @@ class AgentsModule(torch.nn.Module):
         
         x_ego = self.mlp_states_ego(x_ego)
         x_agents = self.mlp_states_agents(x_agents)
-
+        
         # Sum all agents activations in the agents channels per each batch
-        mat = torch.ones(agsshape[0],1,agsshape[1],device=x_agents.device)
+        mat = torch.ones((agsshape[0],1,agsshape[1]),device=x_agents.device, dtype=x_agents.dtype)
         x_agents = torch.bmm(mat,x_agents).squeeze(1)
 
         out = torch.cat([x_ego, x_agents],dim=-1)
